requirements.txt
Flask==2.3.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
